// generated/src/app/api-models/prov-type-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ProvTypeMaster {
  typeOrSpecCode: string;
  typeOrSpecialty: string;
  shortDescription: string;
  description: string;
  autoAuditSpecCod: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}